rename App to resolve casing issues on react update
